backend/app/requirements.txt
fastapi==0.68.2
pydantic==1.9.2
uvicorn==0.15.0
straight.plugin==1.5.0
websockets==10.3
pytest==7.1.2
pytest-cov==3.0.0
python-multipart==0.0.5
python-jose[cryptography]
passlib[bcrypt]
mariadb
